Web-crawler is ready for first release 🏎️🚀
Added the mercator class part of the crawler. Added an example.
Added some tests. and got things ready for first release.

It could totally use better documentation, but this should do for now.

This change includes Prettier changes, so that was quite the mistake 😅
to put them all into one commit, but oh well...